Add support for autocomplete on multi-word commands
Autocomplete previously worked on the whole line instead of singular
words. The preceeding words in a line would therefore be removed if the
user tried to autocomplete on the second (or later) word in the command.

Given a command with multiple words autocomplete now completes each word
rather than the whole line.